Day 5
On a roll, just kept going from Day 4.  This one was trickier, but I was still able to get most of it on my own without really digging too much into google.  I also didn't just convert everything to binary like  I saw a lot of people doing, it's actually sorted.  In otherwords, way more inefficient than the rest, which is the motto of Overengineeing.